Upgrade CodeGenerator to handle new cimgui json files
